emulator file structure reorgnisation and emulator code clarity and error partial improvements

+++ main/ble/gatt_buff.c
@@ -14,7 +14,7 @@ esp_err_t chr_msg_buffer_init(chr_msg_buffer_t *buf)
     return ESP_OK;
 }
 
-esp_err_t chr_msg_buffer_add(chr_msg_buffer_t *buf, const uint8_t *msg, uint16_t len)
+esp_err_t chr_msg_buffer_add(chr_msg_buffer_t *buf, const uint8_t *msg, size_t len)
 {
     if (!buf || !msg || len == 0) return ESP_ERR_INVALID_ARG;
 
@@ -44,7 +44,7 @@ esp_err_t chr_msg_buffer_add(chr_msg_buffer_t *buf, const uint8_t *msg, uint16_t
     return ESP_OK;
 }
 
-esp_err_t chr_msg_buffer_get(chr_msg_buffer_t *buf, size_t index, uint8_t **msg_out, uint16_t *len_out)
+esp_err_t chr_msg_buffer_get(chr_msg_buffer_t *buf, size_t index, uint8_t **msg_out, size_t *len_out)
 {
     if (!buf || !msg_out || index >= buf->count) return ESP_ERR_INVALID_ARG;
